clean up code

+++ longtaskrunnin/interprocess_communication.py
@@ -2,6 +2,7 @@ import pickle
 from pathlib import Path
 from tempfile import mkdtemp
 from typing import Any
+from os import fspath
 
 from balsa import get_logger
 
@@ -20,7 +21,7 @@ class InterprocessCommunication:
 
     def write(self, data: Any):
         """
-        write the result
+        write the data for reading later
 
         :param data: data to write, which will be passed to another process
         """
@@ -28,19 +29,32 @@ class InterprocessCommunication:
             pickle.dump(data, pickle_file)
 
     def get_interprocess_communication_file_path_str(self) -> str:
-        return str(self.interprocess_communication_file_path)
+        """
+        Get the interprocess communication file path as a str
+        :return: interprocess communication file path as a str
+        """
+
+        # Python tip: anytime you accept a path that could be a path-like object (e.g. pathlib), never rely on its string repr; always use os.fsdecode(), os.fsencode(), or os.fspath().
+        # https://twitter.com/brettsky/status/1404521184008413184
+        return fspath(self.interprocess_communication_file_path)
 
 
 def interprocess_communication_read(interprocess_communication_file_path_str: str) -> Any:
     """
     Read the data. Must be called exactly once in order to get the data and clean up temp files.
 
-    :return: data from the .write() call
+    :return: data from the InterprocessCommunication.write() call
     """
+    data = None
     interprocess_communication_file_path = Path(interprocess_communication_file_path_str)
-    assert interprocess_communication_file_path.exists()
-    assert interprocess_communication_file_path.is_file()
-    with open(interprocess_communication_file_path, "rb") as pickle_file:
-        data = pickle.load(pickle_file)
-    rmdir(interprocess_communication_file_path.parent)  # clean up
+
+    # Log errors and return a None instead of taking an exception since PyQt can merely crash on an exception in a thread.
+    if not interprocess_communication_file_path.exists():
+        log.error(f"{interprocess_communication_file_path} does not exist")
+    elif not interprocess_communication_file_path.is_file():
+        log.error(f"{interprocess_communication_file_path} is not a file")
+    else:
+        with open(interprocess_communication_file_path, "rb") as pickle_file:
+            data = pickle.load(pickle_file)
+        rmdir(interprocess_communication_file_path.parent)  # clean up
     return data
